requirements.txt
Flask==3.1.3
python-dotenv==1.2.2
anthropic==0.34.0